+++ .eslintrc.cjs
@@ -3,4 +3,8 @@ module.exports = {
   root: true,
   extends: ["@tailwind-md"],
   ignorePatterns: ["**/**.cjs"],
+  parserOptions: {
+    sourceType: "module",
+    ecmaVersion: "latest",
+  },
 };