+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailwind-md/eslint-config
 
+## 0.0.3
+
+### Patch Changes
+
+- Added state modifier for plugin
+
 ## 0.0.2
 
 ### Patch Changes